fixed warning about loss with size_t to int conversion

+++ ELMS/base_station/base_unit.h
@@ -78,7 +78,7 @@ public:
      * other vehicles */
     void input_data(int index, struct message* ptr, Port& p, HANDLE& h);
     void update_data(struct message* ptr, int indice);
-    int contains_id_number(int id, int& index);
+    int contains_id_number(int id, unsigned long int& index);
     map<int, double> checkDistancesInMasterVector1(Vehicle* v);
     int checkOtherVehiclesPriorityNumbers(Vehicle* v1, int index, int priority_number);
